+++ extensions-imply/imply-timeseries/src/main/java/io/imply/druid/timeseries/aggregation/MeanTimeSeriesAggregatorFactory.java
@@ -76,7 +76,7 @@ public class MeanTimeSeriesAggregatorFactory extends BaseTimeSeriesAggregatorFac
     if (window == null) {
       throw new IAE("Must exclusively have a valid, non-null window");
     }
-    int finalMaxEntries = (int) Math.ceil(window.toDurationMillis() * 1D / timeBucketMillis);
+    int finalMaxEntries = Math.toIntExact((long) Math.ceil(window.toDurationMillis() * 1D / timeBucketMillis));
     if (window.getStartMillis() % timeBucketMillis != 0) {
       finalMaxEntries++;
     }
